Opps Re Pushing
Former-commit-id: 216050ca5580d33fbf411cb00dec00a6f018484c

+++ CometEngine.Win32/CEDESKTOP/com/CometEngine/DeskTop/CEDeskTopMouseMove.java
@@ -7,14 +7,14 @@ import com.CometEngine.Renderer.CERenderer;
 
 public class CEDeskTopMouseMove extends GLFWCursorPosCallback {
 
-	private int CursorPosX = 0;
-	private int CursorPosY = 0;
+	private double CursorPosX = 0;
+	private double CursorPosY = 0;
 
-	public int getXPos() {
+	public double getXPos() {
 		return CursorPosX;
 	}
 
-	public int getYPos() {
+	public double getYPos() {
 		return CursorPosY;
 	}
 
@@ -22,8 +22,8 @@ public class CEDeskTopMouseMove extends GLFWCursorPosCallback {
 	public void invoke(long arg0, double arg1, double arg2) {
 		CERenderer renderer = CometEngine.getInstance().getRenderer();
 
-		CursorPosX = (int) (((double)renderer.getRenderWidth() / (double)CEDeskTop.getFrameWidth()) * arg1);
-		CursorPosY = (int) (((double)renderer.getRenderHeight() / (double)CEDeskTop.getFrameHeight()) * arg2);
+		CursorPosX = (((double)renderer.getRenderWidth() / (double)CEDeskTop.getFrameWidth()) * arg1);
+		CursorPosY =  (((double)renderer.getRenderHeight() / (double)CEDeskTop.getFrameHeight()) * arg2);
 
 		System.out.println("arg0: " + arg0 + " arg1: " + CursorPosX + " arg2: " + CursorPosY);
 	}
